explicitly parse spam_count to integer, to prevent null warning
The spam counter is not initialized by default, so before the very first
spam comment, we load NULL from the database. This triggers a
deprecation warning, because NULL is passed to number_format() in
subsequent calls. Explicitly parse the value as integer yields 0 which
perfectly represents the desired value without any warning.