migrate dashboard chart from Raphael to Chartist (#406)
Use the same chart library and UI as it is used in Statify now.
Doing so, we migrate the graphics library from Raphaël to Chartist with
a tooltips plugin and adjust the wrapper script and styling accordingly.

Including CSS and tooltips plugin, Chartist is about 50KB of size
compared to 96KB for Raphaël.