only determine client IP from trusted headers
By default, we only look for the client's IP address in the REMOTE_ADDR
header. This can be extended via hook if there is a need to use headers
like X-Forwarded-For ("HTTP_X_FORWARDED_FOR") instead.

Headers will be processed in the provided order and return the first
valid IP address found.