Remove `load_plugin_lang()` function and utilize `_load_textdomain_just_in_time()`
This function is calling the `load_plugin_textdomain()`, which eventually become
deprecated soon. Since this plugin is also using the "Text Domain" plugin header,
it is not necessary to call this function at all.

By removing this function, the `_load_textdomain_just_in_time()` function will be
used, when the first string needs to be translated.